requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
dnspython
scikit-learn
shap
anthropic
beautifulsoup4
python-whois
requests
streamlit
mlflow==2.19.0
pyaml
dagshub
fastapi
uvicorn
python-multipart
pymongo[srv]==3.12
dill

-e .